# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing tourist places in Taraz",
        "title": "Taraz Places API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/places": {
            "get": {
                "description": "Returns a list of all places",
                "produces": [
                    "application/json"
                ],
                "summary": "Get all places",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Place"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Adds a new place to the list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create a new place",
                "parameters": [
                    {
                        "description": "Place data",
                        "name": "place",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Place"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Place added",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/places/{id}": {
            "get": {
                "description": "Returns a place by its ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get place by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Place ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Place"
                        }
                    },
                    "404": {
                        "description": "Place not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates an existing place by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update a place",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Place ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated place data",
                        "name": "place",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Place"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Place updated",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a place by ID",
                "summary": "Delete a place",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Place ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Place deleted",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/places/{id}/comment": {
            "post": {
                "description": "Adds a comment to a place",
                "summary": "Add comment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Place ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Comment text",
                        "name": "comment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comment added",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/places/{id}/rate": {
            "patch": {
                "description": "Updates the rating of a place",
                "summary": "Rate a place",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Place ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "New rating",
                        "name": "rating",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rating updated",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.Place": {
            "type": "object",
            "properties": {
                "comments": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "photoURLs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "placeName": {
                    "type": "string"
                },
                "rating": {
                    "type": "number"
                }
            }
        }
    }
}